moved position tests
